config_generator/requirements.txt
click==7.1.2
deepmerge==1.0.1
exo==0.1.2
more_itertools==8.4.0
PyYAML==6.0
requests==2.26.0
terraform_external_data==1.0.3
